Redid file structure. Added tsx check on build. Made build command cross platform.
