+++ src/components/layout/index.ts
@@ -1,0 +1,5 @@
+export { AppTitle } from "./components/AppTitle";
+export { Section } from "./components/Section";
+export { Footer } from "./components/Footer";
+export { Navbar } from "./components/Navbar";
+export { MainLayout } from "./components/MainLayout";
