--- src/components/sponsors/index.ts
@@ -1,2 +1,0 @@
-export * from "./SponsorCard";
-export * from "./PastSponsorAvatar";
