+++ src/features/contributors/components/CircularIcon.tsx
@@ -10,7 +10,7 @@ type CircularIconProps = {
   rel?: string;
 };
 
-export default function CircularIcon(props: CircularIconProps) {
+export const CircularIcon = (props: CircularIconProps) => {
   const BaseElement = props.element ?? "div";
 
   const class_ = `bg-background-10 shadow-lg rounded-full flex items-center justify-center overflow-hidden ${props.class ?? ""}`;
