+++ src/features/contributors/data/contributorsList.ts
@@ -1,38 +1,6 @@
-export interface ContributorDisplay {
-  name: string;
-  src: string;
-}
+import { Contributor } from "../contributors.types";
 
-export interface Contributor {
-  name: string; // internal name for caching not displayed (display#name for user-facing)
-  display: ContributorDisplay[];
-  roles: ("dev" | "artist" | "community")[];
-  color?: string;
-  socials: {
-    twitter?: string;
-    bluesky?: string;
-    github?: string;
-    instagram?: string;
-    youtube?: string;
-    website?: string;
-    twitch?: string;
-    kofi?: string;
-    discord?: string;
-    tiktok?: string;
-    printables?: string;
-    vgen?: string;
-    booth?: string;
-    soundcloud?: string;
-    steam?: string;
-    matrix?: string;
-    reddit?: string;
-    patreon?: string;
-  };
-  tags: string[];
-  classes?: string;
-}
-
-export const contributors: Contributor[] = [
+export const contributorsList: Contributor[] = [
   {
     name: "Butterscotch",
     display: [
