+++ src/features/contributors/index.ts
@@ -1,0 +1,3 @@
+export { Card } from "./components/Card";
+export { CircularIcon } from "./components/CircularIcon";
+export { SearchBox } from "./components/SearchBox";
