+++ src/features/events/index.ts
@@ -1,0 +1,2 @@
+export { EventCard } from "./components/EventCard";
+export { EventsHeader } from "./components/EventsHeader";
