+++ src/features/home/index.ts
@@ -1,0 +1,5 @@
+export { DownloadSection } from "./components/DownloadSection";
+export { HeroSection } from "./components/HeroSection";
+export { QASection } from "./components/QASection";
+export { UseCaseSection } from "./components/UseCaseSection";
+export { VideoSection } from "./components/VideoSection";
