+++ src/features/sponsors/components/PastSponsorAvatar.tsx
