+++ src/features/sponsors/index.ts
@@ -1,0 +1,3 @@
+export { PastSponsorAvatar } from "./components/PastSponsorAvatar";
+export { SponsorCard } from "./components/SponsorCard";
+export type { Sponsor } from "./components/SponsorCard";
