+++ src/routes/events.tsx
@@ -1,21 +1,15 @@
 import { Link, Meta } from "@solidjs/meta";
 import { createMemo, createSignal, For, onMount, Show } from "solid-js";
-import { AppTitle } from "~/components/AppTitle";
-import { Container } from "~/components/commons/Container";
-import { Typography } from "~/components/commons/Typography";
-import { Section } from "~/components/Section";
-import { MainLayout } from "~/layouts/MainLayout";
 import {
   buildEventsJsonLd,
-  getFallbackEvents,
   sortByNextDate,
-  toEventData,
   type EventData,
-} from "~/utils/events";
-import { getCustomEvents } from "~/assets/events-custom";
-import EventsHeader from "~/components/events/EventsHeader";
-import EventCard from "~/components/events/EventCard";
-import { FAQSection } from "~/components/commons/FAQSection";
+} from "~/features/events/utils/events";
+import { fetchEvents } from "~/features/events/api/fetchEventsApi";
+import { EventCard, EventsHeader } from "~/features/events";
+import { getCustomEvents } from "~/features/events/data/events-custom";
+import { Typography, FAQSection, Container } from "~/components/commons";
+import { AppTitle, MainLayout, Section } from "~/components/layout";
 
 const QUESTIONS_COUNT = 5;
 
@@ -23,28 +17,8 @@ export default function EventsPage() {
   const [events, setEvents] = createSignal<EventData[] | null>(null);
 
   onMount(async () => {
-    if (import.meta.env.DEV) {
-      console.log("Development environment, using fallback events");
-      setEvents(getFallbackEvents());
-      return;
-    }
-
-    try {
-      const response = await fetch("/api/events");
-
-      const status = response.status;
-      if (status === 404) console.warn("no events returned");
-      if (status === 500) console.error("server error fetching events");
-      if (!response.ok)
-        throw new Error(`Failed to fetch events: ${response.status}`);
-
-      const result = (await response.json()) as unknown[];
-      setEvents(result.map(toEventData));
-    } catch (error) {
-      const msg = (error as Error).message;
-      console.error(`Failed to fetch events: ${msg}`);
-      setEvents([]);
-    }
+    const fetchedEvents = await fetchEvents();
+    setEvents(fetchedEvents);
   });
 
   const virtualEvents = createMemo(() => sortByNextDate(events() ?? []));
