+++ tsconfig.json
@@ -14,6 +14,7 @@
     "sourceMap": true,
     "paths": {
       "~/*": ["./src/*"]
-    }
+    },
+    "skipLibCheck": true
   }
 }
